Выполнил ДЗ№6. Также, реорганизовал проект согласно предыдущим рекомендациям.
