+++ Swift-learning/Swift-learning/ViewController.swift
@@ -7,161 +7,35 @@
 //
 
 import UIKit
-
-
-enum Level: Int{
-    case full
-    case middle
-    case low
-    case zero
-}
-enum ActionOverACar {
-    case invertEngine
-    case invertWindowsState
-    case loadingStorage(Int)
-    case addRemoveTrailer
-    case putRemoveChildCair
-}
-
-//1. Создать протокол «Car» и описать свойства, общие для автомобилей, а также метод действия.
-protocol Car : CustomStringConvertible{
-    var brand: String { get set }
-    var assembleYear: Int { get set }
-    var storageCapacity: Int { get  set }
-    var storageOccupation: Int { get  set }
-    var isItWounded: Bool { get  set }
-    var windowsAreOpened: Bool { get  set }
-    
-    func performSpecific(action: ActionOverACar) -> Bool
-}
-
-//2.Создать расширения для протокола «Car» и реализовать в них методы конкретных действий с автомобилем:
-//  открыть/закрыть окно, запустить/заглушить двигатель и т.д.
-//  (по одному методу на действие, реализовывать следует только те действия, реализация которых общая для всех автомобилей).
-extension Car {
-    func performBool (action: ActionOverACar) -> Bool {
-        switch action{
-        case .invertEngine:
-            return !isItWounded
-        case .invertWindowsState:
-            return !windowsAreOpened
-        default:
-            print("this action unallowable")
-        }
-        return true
-    }
-    
-    func performLoadUnload (action: ActionOverACar) -> Int{
-        switch action{
-        case .loadingStorage(let load):
-            if 0 <= (storageOccupation + load) && (storageOccupation + load) < storageCapacity {
-                return storageOccupation + load
-            } else if 0 < (storageOccupation + load){
-                return 0
-            } else {
-                return storageCapacity
-            }
-        default:
-            print("this action unallowable")
-        }
-        return 0
-    }
-    
-}
-
 class ViewController: UIViewController {
-  
-// 3. Создать два класса, имплементирующих протокол «Car» - trunkCar и sportСar. Описать в них свойства, отличающиеся для автомобиля и грузовика.
-// 4. Для каждого класса написать расширение, имплементирующее протокол CustomStringConvertible.
-    class PassengerCar: Car{
-        var description: String {
-            return String("-------------------\n" + "brand: \(self.brand)\n" + "assembleYear: \(self.assembleYear)\n" + "storageCapacity: \(self.storageCapacity)\n" + "storageOccupation: \(self.storageOccupation)\n" + "isItWounded: \(self.isItWounded)\n" + "windowsAreOpened: \(self.windowsAreOpened)\n" + "isChildChareExist: \(self.isChildChareExist)\n")
-        }
+    
+//For HW#6
+    @IBAction func checkingHW6(_ sender: UIButton) {
+        var queue = CustomQueue<Int>()
+        queue.push(newItem: 5)
+        queue.push(newItem: 9)
+        queue.push(newItem: 3)
+        queue.push(newItem: 8)
+        queue.push(newItem: 3)
+        queue.push(newItem: 2)
+        queue.push(newItem: 9)
+        queue.push(newItem: 9)
+        print(queue)
         
-        func performSpecific(action: ActionOverACar) -> Bool{
-            switch action{
-            case .putRemoveChildCair:
-                return !self.isChildChareExist
-            default: return !self.isChildChareExist
-            }
-        }
+        let firstInQueue = queue.pull()
+        print("Забрали один элемент: \(firstInQueue!)")
+        print(queue)
         
-        var brand: String
-        var assembleYear: Int
-        var storageCapacity: Int
-        var storageOccupation: Int
-        var isItWounded: Bool
-        var windowsAreOpened: Bool
-        var  isChildChareExist = false
+        print("Все элементы меньше 5:  \(queue.filter(byClosure: {$0<5}))")
+        queue.changeAllItems(byFormula: {$0 + 1})
+        print("Прибавили ко всем элементам еденицу: \n\(queue)")
+        print("Выведем элемент в очереди под номером индексом 3:\n\(queue[3]!)")
+        print("Выведем элемент в очереди под номером индексом 15(По заданию дожно выдать NIL):\n\(String(describing: queue[15]))")
         
-        init(brand: String,
-             assembleYear: Int,
-             storageCapacity: Int,
-             storageOccupation: Int,
-             isItWounded:Bool,
-             windowsAreOpened: Bool,
-             isChildChareExist: Bool){
-            
-            self.brand = brand
-            self.assembleYear = assembleYear
-            self.storageCapacity = storageCapacity
-            self.isItWounded = isItWounded
-            self.storageOccupation = storageOccupation
-            self.windowsAreOpened = windowsAreOpened
-            self.isChildChareExist = isChildChareExist
-            
-        }
     }
     
-    class Lorry: Car {
-        var description: String {
-            return String("-------------------\n" + "brand: \(self.brand)\n" + "assembleYear: \(self.assembleYear)\n" + "storageCapacity: \(self.storageCapacity)\n" + "storageOccupation: \(self.storageOccupation)\n" + "isItWounded: \(self.isItWounded)\n" + "windowsAreOpened: \(self.windowsAreOpened)\n" + "isTrailerExist: \(self.isTrailerExist)\n")
-        }
-        
-        func performSpecific(action: ActionOverACar) -> Bool{
-            switch action{
-            case .addRemoveTrailer:
-                return !self.isTrailerExist
-            default: return self.isTrailerExist
-            }
-        }
-        
-
-        var brand: String
-        var assembleYear: Int
-        var storageCapacity: Int
-        var storageOccupation: Int
-        var isItWounded: Bool
-        var windowsAreOpened: Bool
-        var isTrailerExist = false
-        
-        func performAdoptTrailer(action: ActionOverACar) {
-            
-        }
-        
-        init(brand: String,
-             assembleYear: Int,
-             storageCapacity: Int,
-             storageOccupation: Int,
-             isItWounded:Bool,
-             windowsAreOpened: Bool,
-             isTrailerExist: Bool){
-            self.brand = brand
-            self.assembleYear = assembleYear
-            self.storageCapacity = storageCapacity
-            self.storageOccupation = storageOccupation
-            self.isItWounded = isItWounded
-            self.windowsAreOpened = windowsAreOpened
-            self.isTrailerExist = isTrailerExist
-        }
-    }
-    
- 
-    
-// executable code ----------------------------------------------------------------------------
-    
+//For HW#5
     var vehicle: Car = Lorry(brand: "Noname", assembleYear: 1, storageCapacity: 1, storageOccupation: 1, isItWounded: true, windowsAreOpened: true, isTrailerExist: true)
-
     
     @IBAction func initACar(_ sender: UIButton) {
         
@@ -204,4 +78,14 @@ class ViewController: UIViewController {
             print(vehicle as! PassengerCar)
         }
     }
+    
+    
+
+    
+    
+    
+    
+
+    
+    
 }
